requirements.txt
# general modules
django==3.2
python-decouple==3.8
pillow
djangorestframework

# third party modules
markdown 
django-filter
drf-yasg[validation]
rest_framework_simplejwt

# email thirdparty modules
django-mail-templated

# refactoring and tests
black
flake8
pytest==7.4.0
pytest-django==4.5.2